+++ src/simif.c
@@ -49,7 +49,7 @@ int dispatch_op(void *ud, int op, int32_t addr, int32_t *data)
         // Unavailable memory returns all ones by architectural decision
         // TODO document this in the machine model
         if (op != OP_WRITE)
-            *data = 0xffffffff;
+            *data = (int32_t)0xffffffff;
 
         if (s->conf.verbose > 2) {
             fprintf(stderr, "%-5s @ 0x%08x %s due to invalid device\n",
